snowflake p9

+++ apps/blockml/src/models/timeframe/make-timeframe-day-of-month.ts
@@ -25,7 +25,7 @@ export function makeTimeframeDayOfMonth(item: {
     }
 
     case common.ConnectionTypeEnum.SnowFlake: {
-      sql = `EXTRACT(DAY FROM ${sqlTimestamp})::integer`;
+      sql = `EXTRACT(DAY FROM ${sqlTimestamp})`;
       break;
     }
   }
